pyproject.toml
[project]
name = "coinbase-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.0",
    "apache-airflow-providers-apache-spark>=5.3.2",
    "apache-airflow-providers-fab>=3.0.0",
    "apache-airflow-providers-postgres>=6.3.0",
    "coinbase-advanced-py>=1.8.2",
    "graphviz>=0.21",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pendulum>=3.1.0",
    "pyarrow>=21.0.0",
    "ta-lib>=0.6.7",
]
